Refactor code for Elasticsearch 7
